// src/app/modules/admin/models/assessment.model.ts
export interface Assessment {
  id: number;
  // "createdBy": "admin",
  // "createdAt": "2025-01-02T11:58:55.447",
  // "updatedBy": "Admin",
  // "updatedAt": "2025-01-02T17:24:24.04",
  name: string;
  description: string;
  statusId: number;
  startDateTime: string;
  endDateTime: string;
  isActive: boolean;
}